Experimenting with tensor addition implementation
Going to need to extract tensor access iterator helpers first, but general idea of create a tensor access from memory order of left and attempt to create a tensor access with that same order of right seems to be suitable - means no duplicated 'is this a similar shape' checks. At some point the 'is this a similar shape' checks are going to need to be available as public functions but they should be able to use the same implementation. Once elementwise iteration of a tensor access has a real API, tensor elementwise addition and subtraction should come together nicely. Then just need to document that tensor operations are defind on a 'similar shape' basis, such that as tensor of x: 5, y: 6 can be added to a tensor of y: 6, x: 5 but a tensor of a: 5, b: 6 cannot be added to a tensor of x: 5, y: 6 because the dimensions must match on a name basis, not order.