Rework tensor data_layout again so this time it's a lot easier to reason about
With data_layout explicitly returning Dimension names from the view_shape, TensorAccess becomes a no op, and using the output from data_layout is obviously/self evidently the input to a new TensorAccess to change the view shape to match the efficient iteration. With the data_layout refering to dimension names now, TensorTranspose and TensorRename require non trivial implementations, as TensorRename changes what the dimension names are, which needs to be mapped to the source data_layout and TensorTranspose does similar in the way it retains the 'order' of dimension names while actually swapping their lengths (what was 'a' before no longer really refers to 'a' afterwards). Leaving those for a future commit though