Misc tidyup
Removed out of date comment on TensorAccess, added track caller for all possibly panicking elementwise methods, removed note on adding chain method helpers because I don't want to bless a single one of the 4 constructors for the  method helper, and there's not many great alternatives either. They could all have the same name via a trait implementation, but the discovery of that would be no better than manually constructing the TensorChain and not in scope by default, so it's not really an improvement that way.