Experimenting with a more general version of TensorRef and TensorMut
New traits would be a breaking change to apply to existing TensorRef and TensorMut due to introduced lifetime in trait name. Really want to push the lifetime to an associated type but then can't do default of &T and &mut T which is quite annoying. Need to solve this problem somehow to unblock RecordTensor and RecordMatrix wanting to leverage existing view manipulation code but not being able to return reference types easily as their value would ideally be (T, Index)