Update to textplots 0.8, adding color to kmeans examples
Worked around API breakages in https://github.com/loony-bean/textplots-rs/issues/26 by not creating a binding for the Chart where possible, as if the chart is used immediately via `display` the borrow checker is fine with all the arguments being borrowed. For plotting in a loop, have used the workaround mentioned on the issue, repeatedly using let bindings to shadow the data as it is built up to pass to `lineplot` without creating temporaries that don't live long enough.