Add a TensorReverse type for reversing indexes into a tensor
This should also serve as a partial solution to reverse iterating as a view could be constructed then iterated to accomplish the same. However this is more general as it can reverse only some of the dimensions instead of all which implementing DoubleEndedIterator on the iterator types would not be able to do.