Add method for specifying number of elements in start+end of dimension to retain
Wrappers & matrix impl to follow. This can be generalised to D dimensions, but trying to test that implementation was very difficult because the expected output was so hard to intuit. We can't just drop out the middle 2x2 of a 4x4 tensor because that does't leave a valid shape behind, so when we try to we would get unintuitive results like only the corners. This API forcing the dimensionality to 1 should keep it more intuitive and can still be called multiple times in sequence if needed.